+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Configure optional dependency groups for development dependencies. Requires fewer dependencies to be installed when installing the package.
 
+### Added - 2023-02-28
+- Added smile-Id constants
+- Added custom types
+- Added Business verification
+
 ## [2.0.0] - 2022-11-25
 ### Added
 - Add coverage reports.